api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis[hiredis]==5.0.8
pydantic-settings==2.4.0
python-dotenv==1.0.1
structlog==24.4.0
prometheus-fastapi-instrumentator==6.1.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-multipart==0.0.12